requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
python-multipart
opencv-python-headless